LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__HostHttp.https.tentative-expected.txt


PASS __HostHttp: Does not set via DOM 'Secure; Path=/'
PASS __HostHttp: Does not set via DOM with Domain attribute 'Secure; Path=/; Domain=localhost'
FAIL __HostHttp: Does not set via HTTP with 'Secure; Path=/' assert_equals: expected (undefined) undefined but got (string) "0.6895265113599466"
PASS __HostHttp: Set via HTTP with 'Secure; Path=/; httponly'
FAIL __HostHttp: Does not set via HTTP with 'Secure; Path=/cookies/; httponly' assert_equals: expected (undefined) undefined but got (string) "0.3762235175632438"
FAIL __HostHttp: Does not set via HTTP with 'Path=/;' (without Secure) assert_equals: expected (undefined) undefined but got (string) "0.24820532349333657"
FAIL __HostHttp: Does not set via HTTP with Domain attribute 'Secure; Path=/; Domain=localhost' assert_equals: expected (undefined) undefined but got (string) "0.4930416018851804"
